### 2023-03-01 #5031 @britta-wstnr
Hi @juliensiebert and @jwwthu👋

Would you be willing to review this submission:
TSInterpret: A Python package for the interpretability of time series classification
for [The Journal of Open Source Software (JOSS)](https://joss.theoj.org/)?

The review concerns the software [TSInterpret](https://github.com/fzi-forschungszentrum-informatik/TSInterpret) and a [short paper](https://github.com/openjournals/joss-papers/blob/joss.05031/joss.05031/10.21105.joss.05031.pdf) about it. The review happens fully on GitHub.

You can learn more about the reviewing process and our conflict of interest policy in the reviewer guidelines [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

Please let me know if you are available as a reviewer. If yes, I can add you as a reviewer. The review process will start (in a separate GitHub issue) once a sufficient number of reviewers have been found.

Thank you!
Britta Westner
