+++ TSInterpret/InterpretabilityModels/counterfactual/CF.py
@@ -180,7 +180,7 @@ class CF(InstanceBase):
             item = item.reshape(item.shape[0], item.shape[2], item.shape[1])
         # TODO This is new and needs to be testes
         ind = ""
-        print('Item Shape',item.shape[-2])
+        print("Item Shape", item.shape[-2])
         if item.shape[-2] > 1:
 
             res = (item != exp).any(-1)
