+++ TSInterpret/__version__.py
@@ -1,3 +1,3 @@
-VERSION = (0, 1, 0)
+VERSION = (0, 1, 1)
 
 __version__ = ".".join(map(str, VERSION))  # noqa: F401
