### 2023-03-06 #5220 @editorialbot
[REVIEW]: TSInterpret: A Python package for the interpretability of time series classification
**Submitting author:** @JHoelli (<a href="http://orcid.org/0000-0003-4978-3336">Jacqueline Höllig</a>)
**Repository:** https://github.com/fzi-forschungszentrum-informatik/TSInterpret
**Branch with paper.md** (empty if default branch): paper
**Version:** v.0.2.3
