### 2023-03-10 #5220 @britta-wstnr
Hi @jwwthu and @siebert-julien,

wow you are quick! Thanks for working through the checklist so swiftly.

I would like to as you for some additional information:
- can please link any issues or PRs you opened in the repository https://github.com/fzi-forschungszentrum-informatik/TSInterpret ?
- can you please write up any additional feedback you have for the authors (things you think are great, things you think should or could  be improved in either the paper or the software)?
- can you give a recommendation towards publication based on your review?

Thank you very much for your work on this!
