### 2023-03-10 #5220 @ulf1
@JHoelli @kulbachcedric
What are the software contributions of Steffen Thoma who is listed as author?
https://github.com/fzi-forschungszentrum-informatik/TSInterpret/graphs/contributors
