+++ setup.py
@@ -27,29 +27,28 @@ base_packages = [
     # "scikit-surprise==1.1.1",
     "torch",
     "pandas~=1.3.2",
-    "numpy>=1.21.6",
+    "numpy>=1.21.6,< 2.0",
     "tqdm~=4.61.2",
     "h5py==3.7.0", # todo add version
-    "joblib>=1.0.1",
-    "kaggle==1.5.12",
-    "lime==0.2.0.1",
-    "Markdown==3.3.4",
-    "matplotlib>=3.3.4",
+    "joblib>=1.0.1,< 2.0",
+    #"lime==0.2.0.1",
+    "Markdown==3.3.4,< 4.0",
+    "matplotlib>=3.3.4,< 4.0",
     "partd==1.2.0",
     "pytz>=2021.3",
-    "shap>=0.39.0",
-    "tensorflow>=2.9.1",
-    "keras>=2.9.0",
-    "tsfresh>=0.18.0",
-    "tslearn>= 0.5.2",
-    "seaborn>=0.11.2",
-    "scikit_optimize>=0.9.0",
-    "mlrose>=1.3.0",
-    "torchcam>=0.3.1",
-    "tf_explain>=0.3.1",
+    "shap>=0.39.0,< 1.0",
+    "tensorflow>=2.9.1,< 3.0",
+    "keras>=2.9.0,< 3.0",
+    "tsfresh>=0.18.0,< 1.0",
+    "tslearn>= 0.5.2,< 1.0",
+    "seaborn>=0.11.2,< 1.0",
+    "scikit_optimize>=0.9.0,< 1.0",
+    "mlrose>=1.3.0,< 2.0",
+    "torchcam>=0.3.1,< 1.0",
+    "tf_explain>=0.3.1,< 1.0",
     "opencv-python==4.6.0.66",
-    "captum>= 0.5.0",
-    "pyts>=0.12.0",
+    "captum>= 0.5.0,< 1.0",
+    "pyts>=0.12.0,< 1.0",
     "deprecated==1.2.13",
     "pymop",
     "deap", 
@@ -57,28 +56,28 @@ base_packages = [
 ]
 
 dev_packages = base_packages + [
-    "flake8>=5.0.4",
+    "flake8>=5.0.4,< 6.0",
     "isort==5.9.3",
-    "mypy>=0.761",
-    "pre-commit>=2.9.2",
-    "pytest>=4.5.0",
-    "pytest-cov>=2.6.1",
-    "pyupgrade>=3.2.0"
+    "mypy>=0.761,< 1.0",
+    "pre-commit>=2.9.2,< 3.0",
+    "pytest>=4.5.0,< 5.0",
+    "pytest-cov>=2.6.1,< 3.0",
+    "pyupgrade>=3.2.0,< 4.0"
 ]
 
 docs_packages = [
-    "flask>=2.0.2",
-    "ipykernel>=6.9.0",
-    "mike>=0.5.3",
-    "mkdocs>=1.2.3",
-    "mkdocs-awesome-pages-plugin>=2.7.0",
-    "mkdocs-gen-files>=0.3.5",
-    "mkdocs-literate-nav>=0.4.1",
-    "mkdocs-material>=8.1.11",
-    "mkdocstrings[python]>=0.19.0",
-    'pytkdocs[numpy-style]>=0.5.0',
-    "ipython_genutils>=0.1.0",
-    "mkdocs-jupyter>=0.20.0",
+    "flask>=2.0.2,< 3.0",
+    "ipykernel>=6.9.0,< 7.0",
+    "mike>=0.5.3,< 1.0",
+    "mkdocs>=1.2.3,< 2.0",
[…]
