### 2023-03-10 #5220 @JHoelli
> @JHoelli What are the software contributions of Steffen Thoma who is listed as author? https://github.com/fzi-forschungszentrum-informatik/TSInterpret/graphs/contributors

@ulf1 Steffen Thoma advised and helped to conceptualize this work. Further, he gave feedback on the algorithms included, the visualizations used, the architecture, and useability. He is also directly involved in research projects where TSInterpret is developed.
