### 2023-03-10 #5220 @JHoelli
> @JHoelli Can you name other projects, other end-users, or other research papers that used the software package? (Any showcase or demo deployments somewhere?)
@ulf1 Our project is fairly new (Open Source since August 2022), so we have yet to find any research paper that cites our work due to the review cycle of conferences and the research development process. However, we see great citation potential as no other PyPi installable package focuses on the ubiquitous problem time series interpretability and an increasing number of publications addressing the issue in non-unified ways.
This can be underlined from the PyPI download stats (https://pypistats.org/packages/tsinterpret), our first GitHub issues and stars, and the Twitter attention (https://twitter.com/carlcarrie/status/1576580935189147652 ) we received. Currently, we also use TSInterpret in our government-funded research projects. For known end-users (except our research group), we would refer to the authors of our GitHub issuers, contributors, and stars.
