+++ README.md
@@ -48,13 +48,17 @@ import tensorflow as tf
 
 ```
 ### Create Classifcation Model
+This Sections uses a pretrained Classification Model to illustrate the use of our package. For running the example, please clone our repository. The code in this section can also be replaces with your personal classifiaction model written in torch or tensorflow.
 ```python
 
+# Load data.
 dataset='BasicMotions'
 train_x,train_y, test_x, test_y=UCR_UEA_datasets().load_dataset(dataset)
 enc1=pickle.load(open(f'../../ClassificationModels/models/{dataset}/OneHotEncoder.pkl','rb'))
 train_y=enc1.transform(train_y.reshape(-1,1))
 test_y=enc1.transform(test_y.reshape(-1,1))
+
+# Load a model.
 model_to_explain = tf.keras.models.load_model(f'../../ClassificationModels/models/{dataset}/cnn/{dataset}best_model.hdf5')
 ```
 Explain & Visualize Model
