### 2023-03-10 #5220 @ulf1
@JHoelli
Is somewhere documentation how to use the API classes of the packages to build own XAI classes, e.g. how to use `BaseModel` or `InterpretabilityBase`
