+++ setup.py
@@ -25,7 +25,7 @@ REQUIRES_PYTHON = ">=3.6.0"
 base_packages = [
     "scikit-learn==1.0.2",
     # "scikit-surprise==1.1.1",
-    "torch",
+    "torch>=1.13.0,<2.0",
     "pandas~=1.3.2",
     "numpy>=1.21.6,< 2.0",
     "tqdm~=4.61.2",
