+++ .github/workflows/cla.yml
@@ -24,6 +24,7 @@ jobs:
         uses: contributor-assistant/github-action@v2.2.1
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          PERSONAL_ACCESS_TOKEN: ${{ secrets.PERSONAL_ACCESS_TOKEN }}
         with:
           path-to-signatures: 'signatures/version1/cla.json'
           path-to-document: 'https://elinux.org/Developer_Certificate_Of_Origin' # e.g. a CLA or a DCO document
