+++ .github/workflows/cla.yml
@@ -21,10 +21,10 @@ jobs:
     steps:
       - name: "CLA Assistant"
         if: (github.event.comment.body == 'recheck' || github.event.comment.body == 'I have read the CLA Document and I hereby sign the CLA') || github.event_name == 'pull_request_target'
-        uses: contributor-assistant/github-action@v2.2.1
+        uses: contributor-assistant/github-action@v2.3.0
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          PERSONAL_ACCESS_TOKEN: ${{ secrets.PERSONAL_ACCESS_TOKEN }}
+
         with:
           path-to-signatures: 'signatures/version1/cla.json'
           path-to-document: 'https://elinux.org/Developer_Certificate_Of_Origin' # e.g. a CLA or a DCO document
