### 2023-03-13 #5220 @JHoelli
> @JHoelli Is somewhere documentation how to use the API classes of the packages to build own XAI classes, e.g. how to use `BaseModel` or `InterpretabilityBase`

@ulf1 I added that to the https://github.com/fzi-forschungszentrum-informatik/TSInterpret/blob/main/CONTRIBUTING.md - Section: Adding a new Algorithm.

Thanks for your feedback.
