### 2023-03-13 b8ce69b
Update CONTRIBUTING.md
Typo
