+++ paper/paper.md
@@ -44,7 +44,7 @@ LEFTIST by @guilleme_agnostic_2019 calculates feature importance based on a vari
 
 ![Architecture of TSInterpret.\label{fig:Architecture}](Taxonomy.png){ width=50% }
 
-`TSInterpret` implements these algorithms according to the taxonomy shown in \autoref{fig:Architecture}. The interpretability methods are sorted according to a) the model output (e.g., is a feature map returned or an example time series) and b) the used mechanism (e.g., based on gradients). All implemented objects share a consistent interface to ensure that all methods contain an explain and a plot function. The plot function is implemented on the level below the interface based on the output structure provided by the interpretability algorithm to provide a unified visualization experience (e.g., in the case of Feature Attribution, the plot function visualizes a heatmap on the original sample). If necessary, those plots are refined by the Mechanism layer. The explain function is implemented on the method level. This high reusability ensures the consistency and extensibility of the framework.
+`TSInterpret` implements these algorithms according to the taxonomy shown in \autoref{fig:Architecture}. The interpretability methods are sorted according to a) the model output (e.g., is a feature map returned or an example time series) and b) the used mechanism (e.g., based on gradients). All implemented objects share the interface `InterpretabilityBase` to enforce unified access (i.e., all methods must implement the functions `explain` and `plot`). The `plot` function is implemented on the level below the interface based on the output structure provided by the interpretability algorithm to provide a unified visualization experience (e.g., in the case of Feature Attribution, the `plot` function visualizes a heatmap on the original sample). If necessary, those plots are refined by the Mechanism layer. The `explain` function is implemented on the method level. This high reusability ensures the consistency and extensibility of the framework.
 
 # Acknowledgements
 
