+++ library.bib
@@ -101,3 +101,11 @@ If you use tf-explain in your research, please cite it using these metadata.},
   journal={arXiv preprint arXiv:2104.00950},
   year={2021}
 }
+@inproceedings{hollig2022tsevo,
+  title={TSEvo: Evolutionary Counterfactual Explanations for Time Series Classification},
+  author={H{\"o}llig, Jacqueline and Kulbach, Cedric and Thoma, Steffen},
+  booktitle={2022 21st IEEE International Conference on Machine Learning and Applications (ICMLA)},
+  pages={29--36},
+  year={2022},
+  organization={IEEE}
+}
