+++ paper/paper.bib
@@ -213,3 +213,12 @@ If you use tf-explain in your research, please cite it using these metadata.},
 	pages = {67027--67040},
 	file = {Siddiqui et al. - 2019 - TSViz Demystification of Deep Learning Models for.pdf:C\:\\Users\\nd942\\Zotero\\storage\\VFD5WXLH\\Siddiqui et al. - 2019 - TSViz Demystification of Deep Learning Models for.pdf:application/pdf},
 }
+
+@inproceedings{hollig2022tsevo,
+  title={TSEvo: Evolutionary Counterfactual Explanations for Time Series Classification},
+  author={H{\"o}llig, Jacqueline and Kulbach, Cedric and Thoma, Steffen},
+  booktitle={2022 21st IEEE International Conference on Machine Learning and Applications (ICMLA)},
+  pages={29--36},
+  year={2022},
+  organization={IEEE}
+}
