### 2023-04-03 #5220 @JHoelli
Hi @britta-wstnr ,

Thanks again for your feedback. I accepted your PR went through your comments:

- The image is enlarged
- The description of the repo is switched to english
- I also made some Clarification to Line 52, I hope it is better now
- Both, package and function names are now in machine font
- I also added a reference to TSEvo that is now included too
- Further we went reread the article.

Thanks a lot.
