### 2023-04-11 #5220 @britta-wstnr
Hi @JHoelli
Thanks for adapting the paper. Browsing through it again, I found two morel little nitpicks:
- line 8: typo "unintuitve"
- line 9: something went wrong with the "e.g." in the chain of references (sorry if that was my doing in the PR!)

Furthermore, you can see that the editorialbot reports 8 DOIs as missing: https://github.com/openjournals/joss-reviews/issues/5220#issuecomment-1503498002. Could you please check if they match / find the matching ones and add them to your references?

Thanks! 🙏
