+++ paper/paper.bib
@@ -13,7 +13,7 @@
 @incollection{susto_time-series_2018,
 	title = {Time-{Series} {Classification} {Methods}: {Review} and {Applications} to {Power} {Systems} {Data}},
 	shorttitle = {Time-{Series} {Classification} {Methods}},
-	
+	dot={10.1016/b978-0-12-811968-6.00009-7},
 	language = {en},
 	urldate = {2022-07-14},
 	booktitle = {Big {Data} {Application} in {Power} {Systems}},
@@ -26,6 +26,7 @@
 @article{bagnall_great_2017,
 	title = {The great time series classification bake off: a review and experimental evaluation of recent algorithmic advances},
 	volume = {31},
+	doi={10.1007/s10618-016-0483-9},
 	shorttitle = {The great time series classification bake off},
 	language = {en},
 	number = {3},
@@ -52,6 +53,7 @@
 @article{yang_10_2006,
 	title = {10 {Challenging} {Problems} {In} {Data} {Mining} {Research}},
 	volume = {05},
+	dor={10.1142/s0219622006002258},
 	abstract = {In October 2005, we took an initiative to identify 10 challenging problems in data mining research, by consulting some of the most active researchers in data mining and machine learning for their opinions on what are considered important and worthy topics for future research in data mining. We hope their insights will inspire new research efforts, and give young researchers (including PhD students) a high-level guideline as to where the hot problems are located in data mining.
             Due to the limited amount of time, we were only able to send out our survey requests to the organizers of the IEEE ICDM and ACM KDD conferences, and we received an overwhelming response. We are very grateful for the contributions provided by these researchers despite their busy schedules. This short article serves to summarize the 10 most challenging problems of the 14 responses we have received from this survey. The order of the listing does not reflect their level of importance.},
 	language = {en},
@@ -102,6 +104,7 @@
 @article{ruiz_great_2021,
 	title = {The great multivariate time series classification bake off: a review and experimental evaluation of recent algorithmic advances},
 	volume = {35},
+	doi={10.1007/s10618-020-00727-3},
 	shorttitle = {The great multivariate time series classification bake off},
 	abstract = {Abstract
             Time Series Classification (TSC) involves building predictive models for a discrete target variable from ordered, real valued, attributes. Over recent years, a new set of TSC algorithms have been developed which have made significant improvement over the previous state of the art. The main focus has been on univariate TSC, i.e. the problem where each case has a single series and a class label. In reality, it is more common to encounter multivariate TSC (MTSC) problems where the time series for a single case has multiple dimensions. Despite this, much less consideration has been given to MTSC than the univariate case. The UCR archive has provided a valuable resource for univariate TSC, and the lack of a standard set of test problems may explain why there has been less focus on MTSC. The UEA archive of 30 MTSC problems released in 2018 has made comparison of algorithms easier. We review recently proposed bespoke MTSC algorithms based on deep learning, shapelets and bag of words approaches. If an algorithm cannot naturally handle multivariate data, the simplest approach to adapt a univariate classifier to MTSC is to ensemble it over the multivariate dimensions. We compare the bespoke algorithms to these dimension independent approaches on the 26 of the 30 MTSC archive problems where the data are all of equal length. We demonstrate that four classifiers are significantly more accurate than the benchmark dynamic time warping algorithm and that one of these recently proposed classifiers, ROCKET, achieves significant improvement on the archive datasets in at least an order of magnitude less time than the other three.},
@@ -129,6 +132,7 @@
 	address = {Portland, OR, USA},
 	title = {Agnostic {Local} {Explanation} for {Time} {Series} {Classification}},
 	urldate = {2021-11-10},
[…]
