+++ paper/paper.bib
@@ -13,7 +13,7 @@
 @incollection{susto_time-series_2018,
 	title = {Time-{Series} {Classification} {Methods}: {Review} and {Applications} to {Power} {Systems} {Data}},
 	shorttitle = {Time-{Series} {Classification} {Methods}},
-	dot={10.1016/b978-0-12-811968-6.00009-7},
+	doi={10.1016/b978-0-12-811968-6.00009-7},
 	language = {en},
 	urldate = {2022-07-14},
 	booktitle = {Big {Data} {Application} in {Power} {Systems}},
@@ -53,7 +53,7 @@
 @article{yang_10_2006,
 	title = {10 {Challenging} {Problems} {In} {Data} {Mining} {Research}},
 	volume = {05},
-	dor={10.1142/s0219622006002258},
+	doi={10.1142/s0219622006002258},
 	abstract = {In October 2005, we took an initiative to identify 10 challenging problems in data mining research, by consulting some of the most active researchers in data mining and machine learning for their opinions on what are considered important and worthy topics for future research in data mining. We hope their insights will inspire new research efforts, and give young researchers (including PhD students) a high-level guideline as to where the hot problems are located in data mining.
             Due to the limited amount of time, we were only able to send out our survey requests to the organizers of the IEEE ICDM and ACM KDD conferences, and we received an overwhelming response. We are very grateful for the contributions provided by these researchers despite their busy schedules. This short article serves to summarize the 10 most challenging problems of the 14 responses we have received from this survey. The order of the listing does not reflect their level of importance.},
 	language = {en},
