### 2023-04-11 #5220 @JHoelli
Hi @britta-wstnr ,

thank your for going through the paper again. I fixed both points, added the missing DOI and went through both the paper and the sources again. I also changed the first letters of the title to Upper Case, I hope this is not a problem, I just thought it looks nicer. (Otherwise I can just change it back.)

Thanks
