+++ .github/workflows/unit-tests.yml
@@ -37,7 +37,6 @@ jobs:
           export SKLEARN_ALLOW_DEPRECATED_SKLEARN_PACKAGE_INSTALL=True  
           python -m pip install --upgrade pip
           pip install wheel
-          pip install -e ".[dev]" codecov
       - name: pytest
         run: pytest --cov=. --cov-report=xml
       - name: Upload coverage reports to Codecov with GitHub Action
