+++ .github/workflows/mkdocs.yml
@@ -11,6 +11,6 @@ jobs:
       - uses: actions/setup-python@v2
         with:
           python-version: 3.8
-      - run: export SKLEARN_ALLOW_DEPRECATED_SKLEARN_PACKAGE_INSTALL=True  
+      - run: export SKLEARN_ALLOW_DEPRECATED_SKLEARN_PACKAGE_INSTALL=True 
       - run: pip install .[docs]
       - run: mkdocs gh-deploy --force
