+++ TSInterpret/__version__.py
@@ -1,3 +1,3 @@
-VERSION = (0, 2, 1)
+VERSION = (0, 2, 2)
 
 __version__ = ".".join(map(str, VERSION))  # noqa: F401
