### 2023-04-24 #5220 @JHoelli
Hi @britta-wstnr,

no problem, I hope you feel better.

I tried to fix it, however I could only make it work looking slightly different to the other citations (usually name and year are blue, in the current version only the year is blue).

Do you want me to change all the citations in the Abstract to this style?

If I use [e.g., @kokhlikyan_captum_2020;@klaise_alibi_2021;@meudec_raphael_tf-explain_2021], I get the same issue as before - the e.g. somehow turns up in the middle of the citation. Unfortunatly, I did not find a way around it.

Regarding the points:

- [X] Make a tagged release of your software, and list the version tag of the archived version here. --> v.0.2.3
- [X] Please list the DOI of the archived version here. --> 10.5281/zenodo.7858886

And thanks for asking whether changing capitalization is fine.
