### 2023-04-26 #5220 @britta-wstnr
Hi @JHoelli,

thanks - that looks great now! ✨

One thing I forgot to ask you to do last time (sorry):
- Please carefully double check all author names and their affiliations. I already checked the ORCIDs and they look fine to me, but feel free to double check those as well.

Then I did the last check on the references list in the paper and there is a few things left to fix there (those are now the nitty-gritty details ...):
- _Delaney et al_: The conference name should be capitalized throughout.
- _Fawaz et al_ has a DOI: https://doi.org/10.1007/s10618-019-00619-1, can you please add it?
- _Guilleme et al_: There is a few special characters missing (e.g., Guillemé )
- _Klaise et al_: I could not find a DOI for that paper, but please double check.
- _Meudec_: the Zenodo repository says "tf-explain" (not "Tf-explain")
- _Rajkomar et al_ has a DOI: doi:10.1038/s41746-018-0029-1, can you please add it?
- Please also list the DOIs for all the preprints you cite (additionally to "arXiv:0000:00000"). That would be relevant for _Ismail et al_ and _Kokhlikyan et al_.
