+++ paper/paper.bib
@@ -7,6 +7,7 @@
 	author = {Rajkomar, Alvin and Oren, Eyal and Chen, Kai and Dai, Andrew M. and Hajaj, Nissan and Hardt, Michaela and Liu, Peter J. and Liu, Xiaobing and Marcus, Jake and Sun, Mimi and Sundberg, Patrik and Yee, Hector and Zhang, Kun and Zhang, Yi and Flores, Gerardo and Duggan, Gavin E. and Irvine, Jamie and Le, Quoc and Litsch, Kurt and Mossin, Alexander and Tansuwan, Justin and Wang, De and Wexler, James and Wilson, Jimbo and Ludwig, Dana and Volchenboum, Samuel L. and Chou, Katherine and Pearson, Michael and Madabushi, Srinivasan and Shah, Nigam H. and Butte, Atul J. and Howell, Michael D. and Cui, Claire and Corrado, Greg S. and Dean, Jeffrey},
 	year = {2018},
 	pages = {18},
+	doi={10.1038/s41746-018-0029-1}
 
 }
 
@@ -84,6 +85,7 @@
 	author = {Esling, Philippe and Agon, Carlos},
 	year = {2012},
 	pages = {1--34},
+	
 }
 
 @article{fawaz_deep_2019,
@@ -98,7 +100,8 @@
   number={4},
   pages={917--963},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10618-019-00619-1},
 }
 
 @article{ruiz_great_2021,
