+++ paper/paper.bib
@@ -161,7 +161,7 @@ Section interpretability},
   title={Instance-based counterfactual explanations for time series classification},
   author={Delaney, Eoin and Greene, Derek and Keane, Mark T},
   doi={10.1007/978-3-030-86957-1_3},
-  booktitle={International Conference on Case-Based Reasoning},
+  booktitle={{International} {Conference} on {Case-Based} {Reasoning}},
   pages={32--47},
   year={2021},
   organization={Springer}
