+++ paper/paper.bib
@@ -128,6 +128,7 @@
 	author = {Ismail, Aya Abdelsalam and Gunady, Mohamed and Bravo, Héctor Corrada and Feizi, Soheil},
 	year = {2020},
 	keywords = {Computer Science - Machine Learning, Statistics - Machine Learning},
+	doi={10.48550/arXiv.2010.13924},
 	file = {arXiv Fulltext PDF:C\:\\Users\\nd942\\Zotero\\storage\\E78XLXRN\\Ismail et al. - 2020 - Benchmarking Deep Learning Interpretability in Tim.pdf:application/pdf;arXiv.org Snapshot:C\:\\Users\\nd942\\Zotero\\storage\\HVC8SWGT\\2010.html:text/html},
 }
 
@@ -192,6 +193,7 @@ If you use tf-explain in your research, please cite it using these metadata.},
 	author = {Kokhlikyan, Narine and Miglani, Vivek and Martin, Miguel and Wang, Edward and Alsallakh, Bilal and Reynolds, Jonathan and Melnikov, Alexander and Kliushkina, Natalia and Araya, Carlos and Yan, Siqi and Reblitz-Richardson, Orion},
 	year = {2020},
 	journal = {arXiv:2009.07896},
+	doi={10.48550/arXiv.2009.07896},
 	keywords = {Computer Science - Machine Learning, Statistics - Machine Learning, Computer Science - Artificial Intelligence},
 	file = {arXiv Fulltext PDF:C\:\\Users\\nd942\\Zotero\\storage\\LXVQBS3J\\Kokhlikyan et al. - 2020 - Captum A unified and generic model interpretabili.pdf:application/pdf;arXiv.org Snapshot:C\:\\Users\\nd942\\Zotero\\storage\\E6ZRZDM5\\2009.html:text/html},
 }
