+++ paper/paper.bib
@@ -169,7 +169,7 @@ Section interpretability},
 }
 
 @misc{meudec_raphael_tf-explain_2021,
-	title = {tf-explain},
+	title = {{tf-explain}},
 	copyright = {Open Access},
 	
 	abstract = {Interpretability Methods for tf.keras models with TensorFlow 2.x},
