+++ paper/paper.bib
@@ -139,7 +139,7 @@
 	doi={10.1109/ictai.2019.00067},
 	booktitle = {2019 {IEEE} 31st {International} {Conference} on {Tools} with {Artificial} {Intelligence} ({ICTAI})},
 	publisher = {IEEE},
-	author = {Guilleme, Mael and Masson, Veronique and Roze, Laurence and Termier, Alexandre},
+	author = {Guillem{\'e}, Ma{\"e}l and Masson, V{\'e}ronique and Roz{\'e}, Laurence and Termier, Alexandre},
 	year = {2019},
 	pages = {432--439},
 	annote = {Work section Time-Series Classification
