+++ paper/paper.bib
@@ -207,7 +207,7 @@ If you use tf-explain in your research, please cite it using these metadata.},
 	author = {Klaise, Janis and Looveren, Arnaud Van and Vacanti, Giovanni and Coca, Alexandru},
 	year = {2021},
 	pages = {1--7},
-	doi={10.5555/3546258.3546439}
+
 }
 
 @article{siddiqui_tsviz_2019,
