### 2023-04-26 #5220 @JHoelli
Hi @britta-wstnr,

thanks for rechecking.

- I checked the ORCIDs in both, zenodo and the paper, they are indeed correct.
- I also fixed the remaining sources.
- I also added a DOI for Ates et al.
- For Klaise et al I also failed to find a valid DOI
