### 2023-05-05 #5220 @arfon
Also, I'm delighted to say that this is the 2000th paper published in JOSS, and it's our 7th birthday too 🎉 🎉 🎉  (see https://blog.joss.theoj.org/2023/05/JOSS-publishes-2000th-paper)
